refactor: reorganize image handling components and update package structure
