+++ content/presentation/src/main/java/com/example/content/presentation/details/DetailsScreen.kt
@@ -1,10 +1,7 @@
 package com.example.content.presentation.details
 
-import android.graphics.drawable.BitmapDrawable
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -16,20 +13,15 @@ import androidx.compose.foundation.layout.statusBarsPadding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.asImageBitmap
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.layout.Layout
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
@@ -37,13 +29,10 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import coil.compose.AsyncImagePainter
-import coil.compose.rememberAsyncImagePainter
 import com.example.content.presentation.details.component.paletteBackgroundBrush
 import com.example.content.presentation.details.libs.PokeDicText
 import com.example.content.presentation.details.libs.rememberPaletteState
-import com.example.content.presentation.home.libs.ConstraintsSizeResolver
-import com.example.content.presentation.home.libs.requestOfWithSizeResolver
+import com.example.content.presentation.libs.PokemonImage
 import com.example.core.model.Pokemon
 import com.example.core.model.PokemonDetails
 import com.example.core.presentation.designsystem.JetpackApplicationTheme
@@ -167,7 +156,6 @@ private fun DetailsHeader(
         onPaletteLoaded(palette)
       }
     )
-
   }
 
   PokeDicText(
@@ -180,73 +168,6 @@ private fun DetailsHeader(
     fontWeight = FontWeight.Bold,
     textAlign = TextAlign.Center,
     fontSize = 36.sp
-  )
-}
-
-@Composable
-private fun PokemonImage(
-  imageUrl: String,
-  modifier: Modifier = Modifier,
-  onPaletteLoaded: (Palette) -> Unit,
-) {
-  val request = requestOfWithSizeResolver(
-    model = imageUrl,
-    contentScale = ContentScale.Fit
-  )
-
-  val sizeResolver = request.sizeResolver
-
-  val painter = rememberAsyncImagePainter(model = request)
-
-  val state = painter.state
-  if (state is AsyncImagePainter.State.Success) {
-    val drawable = state.result.drawable
-    if (drawable is BitmapDrawable) {
-      val bitmap = drawable.bitmap
-      val imageBitmap = bitmap.asImageBitmap()
-      LaunchedEffect(imageBitmap) {
-        val palette = Palette.from(imageBitmap).generate()
-        onPaletteLoaded(palette)
-      }
-    }
-  }
-
-  Layout(
-    content = {
-      when (painter.state) {
-        AsyncImagePainter.State.Empty -> Unit
-        is AsyncImagePainter.State.Error -> {
-          Text("Error occurred")
-        }
-
-        is AsyncImagePainter.State.Loading -> {
-          Row(
-            horizontalArrangement = Arrangement.Center,
-            modifier = Modifier.fillMaxWidth()
-          ) {
-            CircularProgressIndicator()
-          }
-        }
-
-        is AsyncImagePainter.State.Success -> {
-          Image(
-            painter = painter,
-            contentDescription = null
-          )
-        }
-      }
-    },
-    modifier = if (sizeResolver is ConstraintsSizeResolver) {
-      modifier.then(sizeResolver)
-    } else {
-      modifier
-    },
-    measurePolicy = { measurable, constraints ->
-      val placeable = measurable.first().measure(constraints)
-      layout(placeable.width, placeable.height) {
-        placeable.placeRelative(0, 0)
-      }
-    }
   )
 }
